+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Added `Set-PnPBuiltInDesignPackageVisibility` cmdlet (0.1.7)
 - Added `Get-PnPExternalUser` cmdlet (0.1.7)
 - Added `Remove-PnPExternalUser` cmdlet (0.1.7)
+- Added `Get-PnPSiteCollectionAppCatalogs` cmdlet (0.1.7)
 
 ### Changed